search関数をモジュールに切り出した
